PKAI Prototype/requirements.txt
streamlit
python-dotenv==0.13
npm
python-docx
IPython
langchain_core
langchain_openai
streamlit_mermaid
llama_index
openpyxl
selenium
streamlit_sortables
streamlit-authenticator
pyodbc
audio_recorder_streamlit
streamlit_float
chromedriver
chromium
nodejs
streamlit-sortables
pyodbc
django-pyodbc-azure
PyPDF2
networkx
call-function-with-timeout